+++ nengo_bones/tests/utils.py
@@ -1,5 +1,6 @@
 """Tools for writing tests in nengo-bones."""
 
+import re
 from traceback import print_tb
 
 
@@ -33,3 +34,33 @@ def assert_exit(result, status):
         print(result.exception)
         print_tb(result.exc_info[2])
         raise
+
+
+def make_has_line(lines, strip=False, regex=False):
+    """Create a function to check file or output lines in order"""
+
+    idx = 0
+
+    def has_line(target, strip=strip, regex=regex, print_on_fail=True):
+        nonlocal lines
+        nonlocal idx
+
+        while idx < len(lines):
+            line = lines[idx]
+            idx += 1
+
+            if strip:
+                line = line.strip()
+
+            match = (re.search(target, line) if regex else
+                     line.startswith(target))
+            if match:
+                return True
+
+        if print_on_fail:
+            print("Failed to find '%s' in\n" % target)
+            print("".join(lines))
+
+        return False
+
+    return has_line